refactor dart int64 write/read
