+++ dart/packages/fory/lib/fory.dart
@@ -30,7 +30,7 @@ export 'src/annotation/fory_struct.dart';
 export 'src/annotation/fory_union.dart';
 export 'src/annotation/type_spec.dart';
 export 'src/annotation/numeric_types.dart';
-export 'src/buffer.dart'
+export 'src/memory/buffer.dart'
     hide
         bufferByteData,
         bufferBytes,
