+++ dart/packages/fory/lib/src/context/meta_string_writer.dart
@@ -19,7 +19,7 @@
 
 import 'dart:collection';
 
-import 'package:fory/src/buffer.dart';
+import 'package:fory/src/memory/buffer.dart';
 import 'package:fory/src/meta/meta_string.dart';
 
 /// Write-side state for meta-string references in one serialization stream.
