+++ dart/packages/fory/lib/src/context/ref_reader.dart
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import 'package:fory/src/buffer.dart';
+import 'package:fory/src/memory/buffer.dart';
 import 'package:fory/src/context/ref_writer.dart';
 
 final class RefReader {
